Refactor RALBackendHandlerFactory and move package of RALUpdater.
