+++ distsql/handler/src/main/java/org/apache/shardingsphere/distsql/handler/ral/update/GlobalRuleRALUpdater.java
@@ -15,28 +15,24 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.distsql.handler.update;
+package org.apache.shardingsphere.distsql.handler.ral.update;
 
+import org.apache.shardingsphere.infra.metadata.ShardingSphereMetaData;
 import org.apache.shardingsphere.infra.util.spi.annotation.SingletonSPI;
 import org.apache.shardingsphere.infra.util.spi.type.typed.TypedSPI;
 import org.apache.shardingsphere.sql.parser.sql.common.statement.SQLStatement;
 
-import java.sql.SQLException;
-
 /**
- * RAL updater.
- * 
- * @param <T> type of updatable RAL statement
+ * RAL updater for global rule.
  */
 @SingletonSPI
-public interface RALUpdater<T extends SQLStatement> extends TypedSPI {
+public interface GlobalRuleRALUpdater extends TypedSPI {
     
     /**
      * Execute update.
      *
-     * @param databaseName database name
-     * @param sqlStatement updatable RAL statement
-     * @throws SQLException SQL exception
+     * @param metaData meta data
+     * @param sqlStatement SQL statement
      */
-    void executeUpdate(String databaseName, T sqlStatement) throws SQLException;
+    void executeUpdate(ShardingSphereMetaData metaData, SQLStatement sqlStatement);
 }
